code's been updated

+++ flask_input_exercise/app.py
@@ -11,10 +11,10 @@ app = Flask(__name__)
 def interest_calculator():
     # handle the POST request = receber
     if request.method == 'POST':
-        principal = int(request.form.get('principal'))
-        interest = int(request.form.get('interest'))/100
-        time = int(request.form.get('time'))
-        total = int(principal * ((1 + interest) ** time))
+        principal = float(request.form.get('principal'))
+        interest = float(request.form.get('interest'))/100
+        time = float(request.form.get('time'))
+        total = float(principal * ((1 + interest) ** time))
         return '''
                 
                   <h1>The amount of money accumulated in {} year(s) is: R$ {:.2f}</h1>
@@ -23,9 +23,9 @@ def interest_calculator():
     # otherwise handle the GET request
     return '''
            <form method="POST">
-               <div><label>Principal:       <input type="int" name="principal">         </label></div>
-               <div><label>Annual Interest: <input type="int" name="interest">          </label></div>
-               <div><label>Time:            <input type="int" name="time">              </label></div>
+               <div><label>Principal:       <input type="float" name="principal">         </label></div>
+               <div><label>Annual Interest: <input type="float" name="interest">          </label></div>
+               <div><label>Time:            <input type="float" name="time">              </label></div>
                <input type="submit" value="Calculate">
            </form>'''
